+++ src/projects/forms/frm_build.py
@@ -18,8 +18,8 @@ FRAME_TITLE = "Build package"
 
 
 class BuildFrame:
-    def __init__(self, parent, project, git_commit: bool):
-        self.root = tk.Toplevel(parent)
+    def __init__(self, parent, project, git_commit: bool) -> None:
+        self.root = tk.Toplevel(parent.root)
         self.project = project
         self.git_commit = git_commit